订单详情显示拼团抽奖信息

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/wxapp/goods/groupDraw/GroupDrawInfoByOsVo.java
@@ -11,7 +11,7 @@ import lombok.Data;
 public class GroupDrawInfoByOsVo {
 	private Integer groupId;
 	private Integer activityId;
-	private Integer activityType;
+	private Byte activityType;
 	private Byte status;
 	private Byte drawStatus;
 	private Byte isGrouper;
